requirements-demo.txt
# Demo app dependencies
streamlit>=1.28.0
streamlit-drawable-canvas>=0.9.3
pillow>=9.0.0
numpy>=1.20.0
torch>=1.9.0
torchvision>=0.10.0
matplotlib>=3.5.0
